clkmgr: Fix clock offset and notification timestamp update
 - Change notification timestamp type to uint64_t representing ns.
 - When using the C API, update client state after wait whether or not an
   event occurs.
 - update notification timestamp printing to keep it as second.
 - update notification timestamp printing to second with 3 decimal places.

+++ pub/c/clkmgr_client_api_c.h
@@ -55,7 +55,7 @@ struct clkmgr_c_state {
     bool gm_changed;                      /**< Primary clock UUID changed */
     bool composite_event;                 /**< Composite event */
     int64_t  clock_offset;                /**< Clock offset */
-    double notification_timestamp;        /**< Timestamp for last notification */
+    uint64_t notification_timestamp;      /**< Timestamp for last notification */
 };
 
 /** Event count for the events */
