+++ pub/clkmgr_client_state.hpp
@@ -35,7 +35,7 @@ struct clkmgr_state {
     bool     gm_changed; /**< Primary clock UUID changed */
     bool     composite_event; /**< Composite event */
     int64_t  clock_offset; /**< Clock offset */
-    double   notification_timestamp; /**< Timestamp for last notification */
+    uint64_t   notification_timestamp; /**< Timestamp for last notification */
 };
 
 /**
